requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0
requests==2.31.0
httpx==0.27.0

# PDF processing
pymupdf==1.23.8
pytesseract==0.3.10
numpy==1.24.3
Pillow==10.0.1
pyvi==0.1.1
huggingface_hub==0.16.4
transformers==4.33.3
pdf2image==1.17.0

# Vector database
qdrant-client==1.6.3
sentence-transformers==2.2.2

# AWS
boto3==1.28.38

# RabbitMQ
pika==1.3.2

# Gmail API
google-api-python-client==2.117.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0

# CUDA 12.1 requirements
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
--find-links https://download.pytorch.org/whl/cu121/torch_stable.html

# LangChain and API requirements
azure-ai-documentintelligence
langchain-core>=0.1.20
langchain-deepseek
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
huggingface_hub>=0.23.0

# Vietnamese NLP requirements
nltk
underthesea
regex
pandas

# Utilities
tqdm==4.66.1
loguru==0.7.0 


